feat: body validation

+++ packages/server/src/server/services/UserService.ts
@@ -1,7 +1,7 @@
 import { UserRepository } from '../repositories'
 import { CreateUserDTO, LoginUserDTO, UpdateUserDTO } from '../models/DTO/UserDTO'
 import { generateToken } from '../libs/jwt'
-
+import { ForbiddenError, NotFoundError } from '../middlewares/HTTPError'
 
 /**
  * The module service implements the business logic for the modules
@@ -10,8 +10,9 @@ export default class UserService {
   constructor(private userRepository: UserRepository) {}
 
   authenticateUser = async (loginUser: LoginUserDTO) => {
-    return this.userRepository.userAuthentification(loginUser)
-      .then(user => {
+    return this.userRepository
+      .userAuthentification(loginUser)
+      .then((user) => {
         // Generate auth token
         const token = generateToken({
           id: user.id,
@@ -42,6 +43,20 @@ export default class UserService {
   }
 
   deleteUser = async (id: number) => {
-    return this.userRepository.deleteUser(id)
+    return this.getUser(id)
+      .then((user) => {
+        if (!user) {
+          throw new NotFoundError('User not found')
+        }
+
+        if (user.isDefault) {
+          throw new ForbiddenError('Cannot delete the default user')
+        }
+
+        return this.userRepository.deleteUser(id)
+      })
+      .catch((err) => {
+        throw err
+      })
   }
 }
